--- src/app/authentication/signup/index.tsx
@@ -1,49 +1,0 @@
-import React from "react";
-
-const SignUpPage = () => {
-  return (
-    <div className="flex items-center justify-center min-h-screen bg-gray-100">
-      <div className="bg-white p-8 rounded-2xl shadow-lg w-full max-w-xl">
-        <h1 className="text-2xl font-bold mb-6 text-center">Sign Up</h1>
-        <div className="grid grid-cols-2 gap-4">
-          {[
-            { label: "Name", name: "name", type: "text" },
-            { label: "Surname", name: "surname", type: "text" },
-            { label: "Username", name: "username", type: "text" },
-            { label: "Email", name: "email", type: "email" },
-            { label: "Password", name: "password", type: "password" },
-            {
-              label: "Confirm Password",
-              name: "confirmPassword",
-              type: "password",
-            },
-          ].map(({ label, name, type }) => (
-            <div className="mb-4" key={name}>
-              <label
-                htmlFor={name}
-                className="block text-sm font-medium text-gray-700"
-              >
-                {label}
-              </label>
-              <input
-                id={name}
-                name={name}
-                type={type}
-                className="mt-1 block w-full px-4 py-2 border border-gray-300 rounded-xl focus:outline-none focus:ring-2 focus:ring-blue-500"
-                required
-              />
-            </div>
-          ))}
-        </div>
-        <button
-          type="submit"
-          className="w-full bg-blue-600 text-white py-2 rounded-xl hover:bg-blue-700 transition"
-        >
-          Create Account
-        </button>
-      </div>
-    </div>
-  );
-};
-
-export default SignUpPage;
